Enumerator support for easy progressbars in ruby >= 2.0
This adds a ProgressBar.iterator method for returning a new Enumerator
that will automatically display a progress bar when enumerating.

If you require 'progressbar/iterator', this will also add a
Enumerator#progressbar method for even easier creation.

For example:

    ProgressBar.iterator(my_array, "Label").each { ... }

With the extension:

    my_array.each.progressbar("Label") { ... }

Also enables travis builds for 2.0.0 and 2.1.0, since otherwise this
wouldn't get tested at all.